Created facade layer fo Activity services
