--- src/main/java/cz/inspire/thesis/data/facade/SportServerFacade.java
@@ -1,4 +1,0 @@
-package cz.inspire.thesis.data.facade;
-
-public class SportServerFacade {
-}
