+++ src/main/java/cz/inspire/thesis/data/service/sport/activity/ActivityService.java
@@ -40,15 +40,9 @@ public class ActivityService {
 
     @Inject
     private ActivityRepository activityRepository;
-    @Inject
-    private InstructorService instructorService;
-    @Inject
-    private SportInstructorService sportInstructorService;
-    @Inject
-    private SportService sportService;
 
     @Transactional
-    public String create(ActivityDetails details) throws CreateException {
+    public ActivityEntity create(ActivityDetails details) throws CreateException {
         try {
             ActivityEntity entity = new ActivityEntity();
             if (details.getId() == null) {
@@ -61,115 +55,21 @@ public class ActivityService {
             entity.setIconId(details.getIconId());
 
             activityRepository.save(entity);
-            postCreate(details, entity);
-            return entity.getId();
+            return entity;
         } catch (Exception e) {
             throw new CreateException("Failed to create Activity entity", e);
         }
     }
 
     @Transactional
-    public void postCreate(ActivityDetails details, ActivityEntity entity) throws CreateException {
-        Set<InstructorDetails> instructors = (Set<InstructorDetails>) details.getInstructors();
-        if (instructors != null && !instructors.isEmpty()) {
-            try {
-                Collection<InstructorEntity> instructorSet = new HashSet<InstructorEntity>();
-                for (InstructorDetails instructorDetails : instructors) {
-                    InstructorEntity instructor = instructorService.findOptionalBy(instructorDetails.getId())
-                            .orElseThrow(() -> new ApplicationException("Instructor entity not found for ID: " + instructorDetails.getId()));
-                    instructorSet.add(instructor);
-                }
-                entity.setInstructors(instructorSet);
-                activityRepository.save(entity);
-            } catch (Exception ex) {
-                throw new CreateException("Failed to associate instructors with activity: " + ex.getMessage(), ex);
-            }
-        }
-    }
-
-    @Transactional
-    public void setDetails(ActivityDetails details) throws ApplicationException {
+    public void saveEntity(ActivityEntity entity) throws ApplicationException {
         try {
-            // Find the activity entity by ID or throw an exception if not found
-            ActivityEntity entity = activityRepository.findOptionalBy(details.getId())
-                    .orElseThrow(() -> new ApplicationException("Activity entity not found for ID: " + details.getId()));
-
-            // Update basic fields
-            entity.setName(details.getName());
-            entity.setDescription(details.getDescription());
-            entity.setIndex(details.getIndex());
-            entity.setIconId(details.getIconId());
-
-            // Fetch current instructor IDs from the entity
-            Set<String> oldInstructorIds = entity.getInstructors().stream()
-                    .map(InstructorEntity::getId)
-                    .collect(Collectors.toSet());
-
-            // Clear existing instructors
-            entity.getInstructors().clear();
-
-            // Add new instructors if provided
-            if (details.getInstructors() != null) {
-                Set<InstructorEntity> instructors = details.getInstructors().stream()
-                        .map(instructorDetails -> {
-                            try {
-                                return instructorService.findOptionalBy(instructorDetails.getId())
-                                        .orElseThrow(() -> new ApplicationException("Instructor not found for ID: " + instructorDetails.getId()));
-                            } catch (Exception e) {
-                                throw new RuntimeException(e);
-                            }
-                        })
-                        .collect(Collectors.toSet());
-
-                entity.setInstructors(instructors);
-            }
-
-            // Save the updated entity
             activityRepository.save(entity);
-
-            // Update related sport-instructor entities
-            updateSportInstructor(details, oldInstructorIds);
-        } catch (Exception ex) {
-            throw new ApplicationException("Failed to update Activity entity: " + ex.getMessage(), ex);
+        } catch (Exception e) {
+            throw new ApplicationException("Failed while saving ActivityEntity : ", e);
         }
     }
 
-    @Transactional
-    private void updateSportInstructor(ActivityDetails activityDetails, Set<String> oldInstructorIds) throws ApplicationException {
-        try {
-            // Extract new instructor IDs from ActivityDetails
-            Set<String> newInstructorIds = activityDetails.getInstructors() != null
-                    ? activityDetails.getInstructors().stream()
-                    .map(InstructorDetails::getId)
-                    .collect(Collectors.toSet())
-                    : new HashSet<>();
-
-            // Determine instructors to delete (oldInstructorIds - newInstructorIds)
-            oldInstructorIds.removeAll(newInstructorIds);
-
-            // Iterate through the sports associated with the activity
-            for (SportDetails sportDetails : activityDetails.getSports()) {
-                // Fetch the corresponding SportEntity
-                SportEntity sportEntity = sportService.findOptionalBy(sportDetails.getId())
-                        .orElseThrow(() -> new ApplicationException("Sport not found for ID: " + sportDetails.getId()));
-
-                // Iterate through sport-instructor mappings
-                for (SportInstructorEntity sportInstructor : sportEntity.getSportInstructors()) {
-                    if (sportInstructor.getInstructor() != null
-                            && oldInstructorIds.contains(sportInstructor.getInstructor().getId())) {
-                        // Mark sport-instructor entity as deleted
-                        sportInstructor.setDeleted(true);
-                        sportInstructorService.setDetails(sportInstructorService.getDetails(sportInstructor)); // Save the updated entity
-                    }
-                }
-
-                // Ensure the sport is not left without any instructors if necessary
-                sportService.checkSportWithoutInstructor(sportEntity);
-            }
-        } catch (Exception ex) {
-            throw new ApplicationException("Failed to update sport-instructor relationships: " + ex.getMessage(), ex);
-        }
-    }
 
     public ActivityDetails getDetails(ActivityEntity entity) {
         ActivityDetails details = new ActivityDetails();
@@ -189,22 +89,17 @@ public class ActivityService {
         return details;
     }
 
-    public Collection<ActivityDetails> findAll() {
-        return activityRepository.findAll().stream()
-                .map(this::getDetails)
-                .collect(Collectors.toList());
+
+    public Collection<ActivityEntity> findAll() {
+        return activityRepository.findAll();
     }
 
-    public Collection<ActivityDetails> findAll(int offset, int count) {
-        return activityRepository.findAll(offset, count).stream()
-                .map(this::getDetails)
-                .collect(Collectors.toList());
+    public Collection<ActivityEntity> findAll(int offset, int count) {
+        return activityRepository.findAll(offset, count);
     }
 
-    public Collection<ActivityDetails> findAllByInstructor(String instructorId, int offset, int count) {
-        return activityRepository.findAllByInstructor(instructorId, offset, count).stream()
-                .map(this::getDetails)
-                .collect(Collectors.toList());
+    public Collection<ActivityEntity> findAllByInstructor(String instructorId, int offset, int count) {
+        return activityRepository.findAllByInstructor(instructorId, offset, count);
     }
 
     public Long countActivities() {
@@ -215,6 +110,13 @@ public class ActivityService {
         return activityRepository.countActivitiesByInstructor(instructorId);
     }
 
+
+    //Additional finders for functionality
+
+    public ActivityEntity findById(String id) throws ApplicationException {
+        return  activityRepository.findOptionalBy(id)
+                .orElseThrow(() -> new ApplicationException("Activity entity not found for ID: " + id));
+    }
     public Optional<ActivityEntity> findOptionalBy(String id) {
         return activityRepository.findOptionalBy(id);
     }
